+++ react/__mocks__/vtex.native-types/index.js
@@ -3,5 +3,13 @@ import { FormattedMessage } from 'react-intl'
 
 export const formatIOMessage = ({ id }) => id
 
-export const IOMessage = props =>
-  props.id ? <FormattedMessage {...props} /> : null
+export const IOMessage = props => {
+  if (props.id) {
+    return <FormattedMessage {...props} />
+  }
+  if (props['data-testid']) {
+    return <div>{props['data-testid']}</div>
+  }
+  return null
+}
+