+++ react/components/SKUSelector/components/Variation.js
@@ -1,5 +1,7 @@
 import PropTypes from 'prop-types'
-import React from 'react'
+import React, { useState, useCallback, memo } from 'react'
+import { Button } from 'vtex.styleguide'
+import { IOMessage } from 'vtex.native-types'
 
 import SelectorItem from './SelectorItem'
 import { stripUrl, isColor } from '../utils'
@@ -8,8 +10,21 @@ import { variationShape } from '../utils/proptypes'
 import styles from '../styles.css'
 import { imageUrlForSize, VARIATION_IMG_SIZE } from '../../module/images'
 
-const Variation = ({ variation, onSelectItem, maxSkuPrice, isSelected }) => {
+const ITEMS_VISIBLE_THRESHOLD = 2
+
+const Variation = ({ variation, onSelectItem, maxSkuPrice, checkSelected, seeMoreLabel, maxItems }) => {
   const displayImage = isColor(variation.name)
+  const { options } = variation
+  const [showAll, setShowAll] = useState(false)
+  const visibleItemsWhenCollapsed = maxItems - ITEMS_VISIBLE_THRESHOLD
+
+  const shouldCollapse = !showAll && options.length > maxItems
+
+  const overflowQuantity = options.length - visibleItemsWhenCollapsed
+
+  const displayOptions = options.slice(0, shouldCollapse ? visibleItemsWhenCollapsed : options.length)
+
+  const showAllAction = useCallback(() => setShowAll(true), [])
 
   return (
     <div
@@ -25,13 +40,13 @@ const Variation = ({ variation, onSelectItem, maxSkuPrice, isSelected }) => {
         >
           {variation.name}
         </span>
-        <div className="inline-flex flex-wrap ml2">
-          {variation.options.map(skuItem => {
+        <div className="inline-flex flex-wrap ml2 flex items-center">
+          {displayOptions.map(skuItem => {
             const [skuImage] = skuItem.images || [null]
             const [seller] = skuItem.sellers
             return (
               <SelectorItem
-                isSelected={isSelected(skuItem)}
+                isSelected={checkSelected(skuItem)}
                 key={skuItem.itemId}
                 isAvailable={seller.commertialOffer.AvailableQuantity > 0}
                 maxPrice={maxSkuPrice}
@@ -54,6 +69,13 @@ const Variation = ({ variation, onSelectItem, maxSkuPrice, isSelected }) => {
               </SelectorItem>
             )
           })}
+          {!showAll && shouldCollapse && (
+            <div className={styles.seeMoreButton}>
+              <Button variation="tertiary" onClick={showAllAction} size="small" collapseLeft>
+                <IOMessage id={seeMoreLabel} values={{ quantity: overflowQuantity }} data-testid="seeMoreLabel" />
+              </Button>
+            </div>
+          )}
         </div>
       </div>
     </div>
@@ -68,7 +90,9 @@ Variation.propTypes = {
   /** Max price of SKU */
   maxSkuPrice: PropTypes.number,
   /** Function to verify if this Variation is selected */
-  isSelected: PropTypes.func,
+  checkSelected: PropTypes.func,
+  seeMoreLabel: PropTypes.string,
+  maxItems: PropTypes.number,
 }
 
-export default Variation
+export default memo(Variation)